폴리곤이 화면에 나온다

+++ src/D3DGraphics/GraphicsEngine.h
@@ -74,8 +74,8 @@ public:
 	void ClearDepthStencilView();
 
 	void CreateInputLayer(ID3D11InputLayout** _inputLayout, ID3D11VertexShader** _vertexShader, ID3D11PixelShader** _pixelShader);
-	void CreateVertexBuffer(Vertex* _verteies, size_t _size, ID3D11Buffer** _vertexbuffer);
-	void CreateIndexBuffer(UINT* _indices, size_t _size, ID3D11Buffer** _indexbuffer);
+	void CreateVertexBuffer(Vertex* _verteies, UINT _size, ID3D11Buffer** _vertexbuffer);
+	void CreateIndexBuffer(UINT* _indices, UINT _size, ID3D11Buffer** _indexbuffer);
 	void CreateRasterizerState(ID3D11RasterizerState** _rasterizerState);
 
 	void SetParameter(DirectX::XMMATRIX _w, DirectX::XMMATRIX _v, DirectX::XMMATRIX _p);
